Convert package objects to toplevel
